Reorganized data handling code
